Rename: move auth files to auth directory
