+++ src/feature/api/auth/auth.module.ts
@@ -1,9 +1,13 @@
+import { MailModule } from '../mail/mail.module';
 import { UsersModule } from '../users/users.module';
-import { AuthController } from './auth.controller';
+import { TwoFactorAuthUseCase } from './application/use-case/two-factor-auth.use-case';
 import { AuthService } from './auth.service';
 import { JwtStrategy } from './jwt/jwt-auth.strategy';
 import { JwtEmailStrategy } from './jwt/jwt-email.strategy';
 import { JwtRegisterStrategy } from './jwt/jwt-register.strategy';
+import { JwtSignInStrategy } from './jwt/jwt-sign-in.strategy';
+import { AuthController } from './presentation/controller/auth.controller';
+import { TwoFactorAuthController } from './presentation/controller/two-factor-auth.controller';
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
@@ -18,8 +22,16 @@ import { JwtModule } from '@nestjs/jwt';
       inject: [ConfigService],
     }),
     UsersModule,
+    MailModule,
   ],
-  controllers: [AuthController],
-  providers: [AuthService, JwtStrategy, JwtEmailStrategy, JwtRegisterStrategy],
+  controllers: [AuthController, TwoFactorAuthController],
+  providers: [
+    AuthService,
+    JwtStrategy,
+    JwtEmailStrategy,
+    JwtRegisterStrategy,
+    JwtSignInStrategy,
+    TwoFactorAuthUseCase,
+  ],
 })
 export class AuthModule {}
