+++ src/feature/api/users/users.module.ts
@@ -4,7 +4,6 @@ import { DeleteFriendUseCase } from './application/friends/delete-friend.use-cas
 import { FindAcceptableFriendRequestUseCase } from './application/friends/find-acceptable-friend-request.use-case';
 import { FindFriendsUseCase } from './application/friends/find-friends.use-case';
 import { FriendRequestUseCase } from './application/friends/friend-request.use-case';
-import { TwoFactorAuthUseCase } from './application/use-case/two-factor-auth.use-case';
 import { UsersUseCase } from './application/use-case/users.use-case';
 import { FriendRequestRepository } from './domain/friend/interface/friend-request.repository';
 import { FriendRepository } from './domain/friend/interface/friend.repository';
@@ -16,7 +15,6 @@ import { FriendRepositoryImpl } from './infrastructure/repositories/friend.repos
 import { UserEntity } from './infrastructure/user.entity';
 import { UserRepositoryImpl } from './infrastructure/user.repository.impl';
 import { FriendsController } from './presentation/controller/friends.controller';
-import { TwoFactorAuthController } from './presentation/controller/two-factor-auth.controller';
 import { UsersController } from './presentation/controller/users.controller';
 import { UsersService } from './users.service';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
@@ -27,13 +25,11 @@ import { Module } from '@nestjs/common';
     MikroOrmModule.forFeature([UserEntity, FriendEntity, FriendRequestEntity]),
     MailModule,
   ],
-  controllers: [UsersController, FriendsController, TwoFactorAuthController],
+  controllers: [UsersController, FriendsController],
   providers: [
     /** application */
     UsersUseCase,
     UsersService,
-
-    TwoFactorAuthUseCase,
 
     FindFriendsUseCase,
     DeleteFriendUseCase,
@@ -57,6 +53,6 @@ import { Module } from '@nestjs/common';
       useClass: FriendRequestRepositoryImpl,
     },
   ],
-  exports: [UsersUseCase, UsersService],
+  exports: [UsersUseCase, UsersService, UserRepository],
 })
 export class UsersModule {}
